requirements.txt
streamlit==0.88.0
face-recognition==1.4.0
opencv-python==4.5.3.56
requests==2.26.0
Pillow==8.3.1
scikit-image==0.18.3